Android 13 (api33) request permission for notification
Maniest permission to post notifications and compose functionality to ask user for notification. TODO: invoke permission request when user first initiates and completes a date choosing their cook schedule. If user does not grant the permission, notify via the UI that they will not be notified (bell with a cross). Otherwise let them know (bell icon) that notifications are active