Feature: Schedule
User selects date and time to eat their pizzas. We keep a schedule displayed for the user to remember their preparation state, and if they want, we also notify them via system notifications
Schedule is persisted so that the user remembers their schedule when they come back to the app.

Doughball preferences (size and amount) are remembered and restored upon application restart.